ui: improve border sizes

+++ src/libtrx/include/libtrx/game/output/draw.h
@@ -36,10 +36,10 @@ void Output_DrawScreenGradientQuad(
 
 void Output_DrawScreenFrame(
     int32_t sx, int32_t sy, int32_t w, int32_t h, RGBA_8888 col_dark,
-    RGBA_8888 col_light, int32_t thickness);
+    RGBA_8888 col_light, float thickness);
 void Output_DrawScreenGradientBox(
     int32_t sx, int32_t sy, int32_t w, int32_t h, RGBA_8888 tl, RGBA_8888 tr,
-    RGBA_8888 bl, RGBA_8888 br, int32_t thickness);
+    RGBA_8888 bl, RGBA_8888 br, float thickness);
 void Output_DrawScreenCentreGradientBox(
     int32_t sx, int32_t sy, int32_t w, int32_t h, RGBA_8888 edge,
-    RGBA_8888 center_h, RGBA_8888 center_v, int32_t thickness);
+    RGBA_8888 center_h, RGBA_8888 center_v, float thickness);
